Additional linking step
When linking initially, questions with identical name and text would not
share a uqid if they had different question number because they would
be matched exactly on question_id. Added a step where the modal
uqid for a question_name, question_text combination is used in place
of less common uqids.